fix: address code review findings
- Fix broken URL templates in fetch_moex.py and fetch_prices.py (double
  braces in non-f-string lines caused .format() to produce literal
  {placeholders} instead of substituted values)
- Change shares_out column from Integer to BigInteger (overflow for
  companies like Sberbank with >2.1B shares)
- Fix admin panel using api_key as session secret_key (should use
  dedicated secret_key for session signing)
- Add SECRET_KEY to docker-compose environment
- Add missing EPS, ROE, dividend_yield mappings to fetch_smartlab.py
  (prevents data loss when re-fetching overwrites previously populated
  fields with None)
- Fix _upsert_reports to skip None values on update (prevents data loss
  when CSV source lacks certain fields)
- Skip migration tests when PostgreSQL is unavailable

+++ investment-api/alembic/versions/85d1cad0af6c_initial.py
@@ -43,7 +43,7 @@ def upgrade() -> None:
     sa.Column('current_price', sa.Numeric(precision=14, scale=2), nullable=True),
     sa.Column('upside', sa.Numeric(precision=8, scale=4), nullable=True),
     sa.Column('market_cap', sa.Numeric(precision=18, scale=2), nullable=True),
-    sa.Column('shares_out', sa.Integer(), nullable=True),
+    sa.Column('shares_out', sa.BigInteger(), nullable=True),
     sa.Column('free_float', sa.Numeric(precision=5, scale=2), nullable=True),
     sa.Column('adv_rub_mln', sa.Numeric(precision=14, scale=2), nullable=True),
     sa.Column('p_e', sa.Numeric(precision=10, scale=2), nullable=True),
